rename search filter -> field filter in filenames
