+++ frontend/src/app/models/index.ts
@@ -1,8 +1,8 @@
 export * from './corpus';
 export * from './found-document';
 export * from './query';
-export * from './search-filter';
-export * from './search-filter-options';
+export * from './field-filter';
+export * from './field-filter-options';
 export * from './base-filter';
 export * from './search-results';
 export * from './sort';
